Source data for Lüders, A., Carpentras, D. and Quayle, M., 2024. Attitude networks as intergroup realities: Using network‐modelling to research attitude‐identity relationships in polarized political contexts. British Journal of Social Psychology, 63(1), pp.37-51.

The sample of N = 402 paid participants through the crowd working platform Prolific Academic. The core of the survey consists of A set of eight political attitude items abortion, immigration, gun control, and gay marriage. Each item followed a 5-point scale ranging from strong disagreement to strong agreement. The survey also includes items on partisanship, affective polarization, and a short vignette experiment.

Usage

data(BrJSocPsychol_2024)

Format

An object of class "data.frame"

References

This dataset was made available by Lüders et.al. 2024.

Examples

data(BrJSocPsychol_2024)
head(BrJSocPsychol_2024)
#> # A tibble: 6 × 234
#>   StartDate  EndDate Progress Duration (in seconds…¹ Finished RecordedDate Q7   
#>   <chr>      <chr>   <chr>    <chr>                  <chr>    <chr>        <chr>
#> 1 44385.536… 44385.… 100      519                    True     44385.54235… Yes  
#> 2 44385.544… 44385.… 100      358                    True     44385.54872… Yes  
#> 3 44385.551… 44385.… 100      524                    True     44385.55780… Yes  
#> 4 44385.557… 44385.… 100      414                    True     44385.56237… Yes  
#> 5 44385.558… 44385.… 100      630                    True     44385.56588… Yes  
#> 6 44385.566… 44385.… 100      265                    True     44385.56952… Yes  
#> # ℹ abbreviated name: ¹​`Duration (in seconds)`
#> # ℹ 227 more variables: Q9_1 <chr>, Q9_2 <chr>, Q9_3 <chr>, Q9_4 <chr>,
#> #   Q9_5 <chr>, Q9_6 <chr>, Q9_7 <chr>, Q9_8 <chr>, Q11 <chr>, Q13 <chr>,
#> #   Q17_1 <chr>, Q17_2 <chr>, Q17_3 <chr>, Q15_1 <chr>, Q15_2 <chr>,
#> #   Q15_3 <chr>, Abort1_dem_1 <chr>, Abort1_rep_1 <chr>, Abort1_neu_1 <chr>,
#> #   Abort1_feel_1 <chr>, Abort1_proto_1 <chr>, Abort2_dem_1 <chr>,
#> #   Abort2_rep_1 <chr>, Abort2_neu_1 <chr>, Abort2_feel_1 <chr>, …